+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This document logs the changes per release of TWCManager.
       * Add a sleep of 5 seconds when waking car up to avoid an infinite loop (thanks @dschuesae)
       * Fix a bug with the legacy web interface which causes the Resume Track Green Energy setting of None to fail. Also added a deprecation notice to the web interface to ensure people don't inadvertently use it over the modular interface.
       * Fixed the Enphase EMS module which was reporting generation values as consumption (thanks @integlikewoah)
+      * Added fix to avoid exception if an incoming TWC message is passed as an immutable bytes object to the unescape_msg function
 
 ## v1.2.0 - 2020-10-09